dev/1.3.0: working on UX

+++ SmartApp/Core/1.Repositories/DataBaseRepository/DataBaseRepository+TrackedEntity.swift
@@ -49,35 +49,42 @@ public extension DataBaseRepository {
         let instances = try? context.fetch(DBEntity.fetchRequestWith(id: trackedEntity.id))
         if let some = instances?.first {
             some.bindWith(model: trackedEntity)
-            // Delete/update current events
-            var currentEventsId: [String] = []
-            some.events?.forEach { event in
-                if let storedEvent = event as? CDataTrackedLog {
-                    let relatedEvent = trackedEntity.cascadeEvents?.filter { $0.id == storedEvent.id }.first
-                    let wasDeleted = relatedEvent == nil
-                    if wasDeleted {
-                        // Deleted
-                        context.delete(event as! NSManagedObject)
-                    } else if let relatedEvent = relatedEvent {
-                        // Updated?
-                        storedEvent.bind(model: relatedEvent)
-                        currentEventsId.append(relatedEvent.id)
+            
+            // Delete/update current events. We ONLY do this, if we have cascadeEvents != nil.
+            // (We could have fetched the `TrackedEntity` without the `cascade` option just to
+            // update one property. Dont remove the type/cast
+            let canUpdateEvents = trackedEntity.cascadeEvents != nil
+            if canUpdateEvents {
+                var currentEventsId: [String] = []
+                some.events?.forEach { event in
+                    if let storedEvent = event as? CDataTrackedLog {
+                        let relatedEvent = trackedEntity.cascadeEvents?.filter { $0.id == storedEvent.id }.first
+                        let wasDeleted = relatedEvent == nil
+                        if wasDeleted {
+                            // Deleted
+                            context.delete(event as! NSManagedObject)
+                        } else if let relatedEvent = relatedEvent {
+                            // Updated?
+                            storedEvent.bind(model: relatedEvent)
+                            currentEventsId.append(relatedEvent.id)
+                        }
                     }
                 }
-            }
-            // Add new events
-            if let cascadeEvents = trackedEntity.cascadeEvents, !cascadeEvents.isEmpty {
                 // Add new events
-                cascadeEvents.forEach { event in
-                    let isNew = !currentEventsId.contains(event.id)
-                    if isNew {
-                        let trackedLog: CDataTrackedLog = CDataTrackedLog(context: context)
-                        trackedLog.id = UUID().uuidString
-                        trackedLog.bind(model: event)
-                        some.addToEvents(trackedLog)
+                if let cascadeEvents = trackedEntity.cascadeEvents, !cascadeEvents.isEmpty {
+                    // Add new events
+                    cascadeEvents.forEach { event in
+                        let isNew = !currentEventsId.contains(event.id)
+                        if isNew {
+                            let trackedLog: CDataTrackedLog = CDataTrackedLog(context: context)
+                            trackedLog.id = UUID().uuidString
+                            trackedLog.bind(model: event)
+                            some.addToEvents(trackedLog)
+                        }
                     }
                 }
             }
+
             CommonCoreData.Utils.save(viewContext: context)
             return trackedEntity.id
         } else {
